+++ rt-in-one-weekend/src/main.cc
@@ -3,6 +3,8 @@
 #include "vec3.h"
 
 #include "useful_utils.h"
+#define STB_IMAGE_WRITE_IMPLEMENTATION
+#include "stb_image_write.h" // using to save PPM to JPG
 
 #include <iostream>
 #include <vector>
@@ -19,12 +21,12 @@ color ray_color(const ray &r)
 int main()
 {
     // Image
+    const int channels_num = 3;
     const auto aspect_ratio = 16.0 / 9.0;
     const int image_width = 400;
     const int image_height = static_cast<int>(image_width / aspect_ratio);
 
     // Camera
-
     auto viewport_height = 2.0;
     auto viewport_width = aspect_ratio * viewport_height;
     auto focal_length = 1.0;
@@ -34,11 +36,11 @@ int main()
     auto vertical = vec3(0, viewport_height, 0);
     auto lower_left_corner = origin - horizontal / 2 - vertical / 2 - vec3(0, 0, focal_length);
 
-    // @TODO: do not forget to rewrite this shit
-    vector<vector<RGB>> pixels_data(image_width, vector<RGB>(image_height)); // Defaults to zero initial value[image_width][image_height];
+    // Store data in std::vector
+    int index = 0;
+    vector<unsigned char> pixels_data(image_width * image_height * channels_num);
 
     // Render
-
     std::cout << "P3\n"
               << image_width << " " << image_height << "\n255\n";
 
@@ -51,16 +53,20 @@ int main()
             auto v = double(j) / (image_height - 1);
             ray r(origin, lower_left_corner + u * horizontal + v * vertical - origin);
             color pixel_color = ray_color(r);
-            //write_color(std::cout, pixel_color);
-            pixels_data[i][j] = write_color_epta(std::cout, pixel_color);
+            write_color(std::cout, pixel_color);
+
+            pixels_data[index++] = static_cast<int>(255.999 * pixel_color.x());
+            pixels_data[index++] = static_cast<int>(255.999 * pixel_color.y());
+            pixels_data[index++] = static_cast<int>(255.999 * pixel_color.z());
         }
     }
 
     // save to .ppm file
-    save_to_ppm((get_time_str() + "_image").c_str(), image_width, image_height, pixels_data);
+    save_to_ppm((get_time_str() + "_blue_grad").c_str(), image_width, image_height, pixels_data);
     // save to .jpg file
-    save_to_jpg(image_width, image_height, pixels_data);
-
+    std::cout << "Trying to write JPG data ........." << std::endl;
+    stbi_write_jpg((get_time_str() + "_blue_grad.jpg").c_str(), image_width, image_height, channels_num, pixels_data.data(), image_width * channels_num);
+    
     std::cerr << "\nDone.\n";
     return 0;
 }
